Add project structure
Add .cobra.yaml for use with cobra-cli

Add serverCmd for basic http API endpoint. The server will crash after 3
seconds if it cannot find a valid mongodb instance specified as a flag
(--mongo-uri) or in config (mongo.uri)

Add api/ package to allow querying the api endpoint through http

Add apiCmd and ping Cmd to query the api for routes and healthcheck,
respectively

Add Dockerfile and docker-compose.yaml for deployment. Will build and
deploy a haul instance, and deploy a mongodb instance only accessible to
it.

Update README to explain how to start the server and access the API.