WIP: Add createComponentCmd
`create component` allows inserting a new component or list of
components into the database.

The component(s) are taken from a list of args, and must parse as valid
JSON dictionaries, which must respect the fields of the types.Component
struct. At least a Component.Name must be provided, but Component.Tags are
non-mandatory.

Currently only Unmarshals and re-Marshals the provided args and print
the resulting objects back to os.Stdout, while printing any errors
encountered to os.Stderr